Migrate backward compatible protocol service
